refactor(runtime): move doctor flavor selection into commands
